+++ pkg/apis/camel/v1/integration_types_support_test.go
@@ -62,7 +62,10 @@ func TestLanguageAlreadySet(t *testing.T) {
 
 func TestAddDependency(t *testing.T) {
 	integration := IntegrationSpec{}
-	integration.AddDependency("camel-file")
+	integration.AddDependency("camel:file")
+	assert.Equal(t, integration.Dependencies, []string{"camel:file"})
+	// adding the same dependency twice won't duplicate it in the list
+	integration.AddDependency("camel:file")
 	assert.Equal(t, integration.Dependencies, []string{"camel:file"})
 
 	integration = IntegrationSpec{}
@@ -72,13 +75,6 @@ func TestAddDependency(t *testing.T) {
 	integration = IntegrationSpec{}
 	integration.AddDependency("file:dep")
 	assert.Equal(t, integration.Dependencies, []string{"file:dep"})
-}
-
-func TestNormalizeDependency(t *testing.T) {
-	assert.Equal(t, "camel:file", NormalizeDependency("camel-file"))
-	assert.Equal(t, "camel:file", NormalizeDependency("camel:file"))
-	assert.Equal(t, "camel:file", NormalizeDependency("camel-quarkus-file"))
-	assert.Equal(t, "camel:file", NormalizeDependency("camel-quarkus:file"))
 }
 
 func TestGetConfigurationProperty(t *testing.T) {
